I logged out all areas the data was being pushed or returned and then went back through and made the console output easier to read.

+++ public/js/controllers.js
@@ -22,7 +22,7 @@ angular.module('mainApp.controllers', [])
   // set session data for page
   $scope.session = Session;
 }])
-.controller('profileController', ['$scope', 'SessionService', 'userService', function($scope, Session, user) {
+.controller('profileController', ['$scope', '$route', 'SessionService', 'userService', function($scope, $route, Session, user) {
   // set session data for page
   $scope.session = Session;
   $scope.isProfile = true;
@@ -60,12 +60,19 @@ angular.module('mainApp.controllers', [])
       $scope.user.RoleId = this.RoleId;
     }
     
+    console.log("The value being sent:");
     console.log($scope.user);
+    console.log("______________");
     // logging the returned values to check if the save was a success
     // Even though a success is returned the values in the objects
     // are not changed
     user.userSaveEdit($scope.user)
-    .then(function(result) {console.log(result);});
+    .then(function(result) {
+      console.log("This is what is returned:");
+      console.log(result);
+      console.log("-----------------------------------");
+    })
+    .then(function() {$route.reload();});
   };
 }])
 .controller('userListController', ['$scope', '$modal', '$route', 'SessionService', 'userService', function($scope, $modal, $route, Session, User){
@@ -83,23 +90,31 @@ angular.module('mainApp.controllers', [])
 
     // edit user
     $scope.editUser = function(user) {
-      $scope.open(user, User);
+      $scope.open(user);
     };
 
     // create a new user
     $scope.createUser = function() {
+      // Sends null so that the modal will know to make a blank user to fill.
       $scope.open(null);
     };
 
     // delete user
     $scope.deleteUser = function(user) {
-      User.deleteUser(user.id)
-      .then(function(data) {
-        if (data !== undefined && data.meta.success) {
-          $route.reload();
-          Session.makeAlert("success","User was successfully deleted");
+      if(user.id != $scope.session.info.id) {
+        if(confirm("Deleting "+user.username+" cannot be undone. Are you sure?"))
+        {
+          User.deleteUser(user.id)
+          .then(function(data) {
+            if (data !== undefined && data.meta.success) {
+              $route.reload();
+              Session.makeAlert("success","User was successfully deleted");
+            }
+          });
         }
-      });
+      } else {
+        alert("You cannot delete yourself!");
+      }
     };
 
     $scope.open = function(user) {
@@ -116,9 +131,19 @@ angular.module('mainApp.controllers', [])
         } else {
           if(result.id) {
             // logging the returned values to check if the save was a success
-            User.userSaveEdit(result).then(function(res) {console.log(res);});
+            User.userSaveEdit(result).then(function(res) {
+              console.log("Edited user from Modal returned info:");
+              console.log(res);
+              console.log("------------------------------------------");
+            });
+            $route.reload(); //To make sure the list is up to date.
           } else {
-            User.createUser(result).then(function(res) {console.log(res);});
+            User.createUser(result).then(function(res) {
+              console.log("Created user from Modal returned info:");
+              console.log(res);
+              console.log("------------------------------------------");
+            });
+            $route.reload(); //To make sure the list is up to date.
           }
         }
       });
@@ -139,7 +164,6 @@ angular.module('mainApp.controllers', [])
         RoleId: null
       };
     }
-    console.log($scope.user);
     $scope.isModal = true;
     $scope.issue = {
       isIssue: false,
@@ -151,6 +175,8 @@ angular.module('mainApp.controllers', [])
       $scope.header = "Create User";
     }
 
+
+    // Username is a required field
     $scope.saveEdit = function () {
       if (this.username) {
         $scope.user.username = this.username;
@@ -162,6 +188,7 @@ angular.module('mainApp.controllers', [])
         }
       }
 
+      // Password is a required field
       if (this.password) {
         if($scope.user.password === null) {
           if (confirm('Are you sure you want a new password?')) {
@@ -181,6 +208,7 @@ angular.module('mainApp.controllers', [])
         }
       }
 
+      // Role Id is a required field
       if (this.RoleId) {
         $scope.user.RoleId = this.RoleId;
       } else { //To ensure new users have a role
@@ -197,11 +225,19 @@ angular.module('mainApp.controllers', [])
       if (this.lastName) {
         $scope.user.lastName = this.lastName;
       }
+      
+
+      // E-Mail is a required field
       if (this.email) {
         $scope.user.email = this.email;
+      } else {
+        if($scope.user.email === null) {
+          $scope.issue.isIssue = true;
+          $scope.issue.problem = "E-Mail";
+          $modalInstance.close($scope.issue);
+        }
       }
 
-      console.log($scope.user);
       $modalInstance.close($scope.user);
     };
